move inv and cred plugins into awx_plugins
