+++ awx_plugins/credentials/injectors.py
